+++ CHANGELOG.md
@@ -46,6 +46,12 @@ Ref: https://keepachangelog.com/en/1.0.0/
 
 ## Unreleased
 
+### Features
+
+### Improvements
+
+- [2328](https://github.com/umee-network/umee/pull/2328) Add UX and `uux` denom aliases for UMEE. 
+
 ## v6.3.0 - 2024-01-03
 
 ### Improvements